requirements本番.txt
Flask==2.2.2
Flask-Cors==4.0.1
aiofiles==24.1.0
aiohttp==3.9.1
anyio==4.2.0
attrs==23.1.0
azure-core==1.30.1
azure-identity==1.17.1
azure-storage-blob==12.16.0
deepgram-sdk==2.12.0
nest-asyncio==1.5.6
openai==0.28.0
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.31.0
typing_extensions==4.9.0
gunicorn==20.1.0
Werkzeug==2.2.3